+++ ios/Classes/MobileScanner.swift
@@ -136,6 +136,15 @@ public class MobileScanner: NSObject, AVCaptureVideoDataOutputSampleBufferDelega
         }
 
         device.addObserver(self, forKeyPath: #keyPath(AVCaptureDevice.torchMode), options: .new, context: nil)
+        do {
+            try device.lockForConfiguration()
+            device.focusMode = AVCaptureDevice.FocusMode.continuousAutoFocus
+            if #available(iOS 15.4, *) {
+                device.automaticallyAdjustsFaceDrivenAutoFocusEnabled = false
+            }
+            device.unlockForConfiguration()
+        } catch {}
+
         captureSession.beginConfiguration()
 
         // Add device input